fix: unread counts not updating when switching courses
Unread counts are unique to each course, and so must not be nested under
the site-wide `session` store.
